fix(web): move tour style constants out of the component file
Exporting TOUR_RUNNER_OPTIONS and TOUR_RUNNER_STYLES from TourRunner.tsx
tripped react-refresh/only-export-components, failing `eslint .`. Relocate
both to a sibling tourRunnerStyles.ts that uses type-only react-joyride
imports, so TourRunner stays the sole runtime importer of the engine and
the lazy-loaded chunk is unaffected.
