+++ web/src/components/tour/TourRunner.theme.test.tsx
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { TOUR_RUNNER_STYLES } from "./TourRunner";
+import { TOUR_RUNNER_STYLES } from "./tourRunnerStyles";
 
 describe("TourRunner theme styles", () => {
   it("uses the resolved text-on-brand token for the primary button", () => {
