+++ web/tests/coverage-matrix.json
@@ -2190,6 +2190,7 @@
       "components": [
         "web/src/components/Dashboard.tsx",
         "web/src/components/tour/TourRunner.tsx",
+        "web/src/components/tour/tourRunnerStyles.ts",
         "web/src/lib/brandMark.ts",
         "web/src/lib/tourSteps.ts",
         "web/src/hooks/useTour.tsx"
